railway_requirements.txt
flask==3.1.2
flask-socketio==5.5.1
python-dotenv==1.1.1
requests==2.32.5
edge-tts==7.2.3
numpy==1.26.4
